Segregates data frame comparisons by type

Splits the output of compare_df_wrapper() in several smaller data frames organized by type of change and formats it

Usage

get_top_bottom_view(
  comparison_df,
  comparison_table_diff,
  id_cols,
  cc_out,
  df1_exact_dups,
  df2_exact_dups,
  df1_id_dups,
  df2_id_dups,
  df1_id_NA,
  df2_id_NA
)

Arguments

comparison_df

the comparison_df element of a compare_df_wrapper() output

comparison_table_diff

the comparison_table_diff element of a compare_df_wrapper() output

id_cols

a character vector of id columns output from compare_cols()

cc_out

a list output from compare_cols()

df1_exact_dups, df2_exact_dups

an exact_dups list element from a find_dups() output, for df1 and df2, respectively

df1_id_dups, df2_id_dups

an id_dups list element from a find_dups() output, for df1 and df2, respectively

df1_id_NA, df2_id_NA

an id_NA list element from a find_dups() output, for df1 and df2, respectively

Value

a data frame with all data from df1 and df2 with comparison annotations where each row in df1 is displayed directly above its df2 counterpart